+++ GUI/Dialogs/ManageGameInstancesDialog.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 using System.IO;
@@ -73,27 +74,15 @@ namespace CKAN
             GameInstancesListView.Items.Clear();
             UpdateButtonState();
 
-            if (!GameInstancesListView.Columns.Contains(GamePlayTime))
-            {
-                // Always show the play time column so our rows load correctly
-                GameInstancesListView.Columns.Insert(
-                    GameInstallPath.Index, GamePlayTime);
-            }
+            var allSameGame = _manager.Instances.Select(i => i.Value.game).Distinct().Count() <= 1;
+            var hasPlayTime = _manager.Instances.Any(instance => (instance.Value.playTime?.Time ?? TimeSpan.Zero) > TimeSpan.Zero);
+
+            AddOrRemoveColumn(GameInstancesListView, Game, !allSameGame);
+            AddOrRemoveColumn(GameInstancesListView, GamePlayTime, hasPlayTime);
 
             GameInstancesListView.Items.AddRange(_manager.Instances
                 .OrderByDescending(instance => instance.Value.Version())
-                .Select(instance => new ListViewItem(new string[]
-                {
-                    !instance.Value.Valid
-                        ? string.Format(Properties.Resources.ManageGameInstancesNameColumnInvalid, instance.Key)
-                        : _manager.CurrentInstance != instance.Value && instance.Value.IsMaybeLocked
-                            ? string.Format(Properties.Resources.ManageGameInstancesNameColumnLocked, instance.Key)
-                            : instance.Key,
-                    instance.Value.game.ShortName,
-                    FormatVersion(instance.Value.Version()),
-                    instance.Value.playTime?.ToString() ?? "",
-                    instance.Value.GameDir().Replace('/', Path.DirectorySeparatorChar)
-                })
+                .Select(instance => new ListViewItem(rowItems(instance.Value, !allSameGame, hasPlayTime))
                 {
                     Tag = instance.Key
                 })
@@ -102,13 +91,41 @@ namespace CKAN
 
             GameInstancesListView.AutoResizeColumns(ColumnHeaderAutoResizeStyle.ColumnContent);
             GameInstancesListView.AutoResizeColumns(ColumnHeaderAutoResizeStyle.HeaderSize);
+        }
 
-            var hasPlayTime = _manager.Instances.Any(instance => (instance.Value.playTime?.Time ?? TimeSpan.Zero) > TimeSpan.Zero);
-            if (!hasPlayTime && GameInstancesListView.Columns.Contains(GamePlayTime))
+        private void AddOrRemoveColumn(ListView listView, ColumnHeader column, bool condition)
+        {
+            if (condition && !listView.Columns.Contains(column))
             {
-                // Hide the play time column if not in use
-                GameInstancesListView.Columns.Remove(GamePlayTime);
+                listView.Columns.Insert(column.Index, column);
             }
+            else if (!condition && listView.Columns.Contains(column))
+            {
+                listView.Columns.Remove(column);
+            }
+        }
+
+        private string[] rowItems(GameInstance instance, bool includeGame, bool includePlayTime)
+        {
+            var list = new List<string>
+            {
+                !instance.Valid
+                    ? string.Format(Properties.Resources.ManageGameInstancesNameColumnInvalid, instance.Name)
+                    : !(_manager.CurrentInstance?.Equals(instance) ?? false) && instance.IsMaybeLocked
+                        ? string.Format(Properties.Resources.ManageGameInstancesNameColumnLocked, instance.Name)
+                        : instance.Name
+            };
+
+            if (includeGame)
+                list.Add(instance.game.ShortName);
+
+            list.Add(FormatVersion(instance.Version()));
+
+            if (includePlayTime)
+                list.Add(instance.playTime?.ToString() ?? "");
+
+            list.Add(instance.GameDir().Replace('/', Path.DirectorySeparatorChar));
+            return list.ToArray();
         }
 
         private static string FormatVersion(GameVersion v)